Split waveform edit and state ownership
